corro-agent: move agent::setup into separate module
